Rename generators to libnamegen
