+++ libnamegen/classic.py
