+++ libnamegen/random.py
